[Refactor]: addbutton 위치 변경
